Added lightning bolts to Zenitsu section will need to adjust so it works with intersection observer tomorrow, and then test to see how both sections are working.
Also, I adjusted the Zenitsu section to have different
height and width dimensions, so check that these also
work with section 1. If so, apply to all sections. I
noticed that the sections seemed to take up almost all
of the screen height on desktop, so adjusting them
slightly down so there's a little bit of padding
will make sure that the animations trigger at the
proper time and that no single part of the section is
out of screen.