JavaScript Zenitsu Animation Fix I decided I didn't like how the lightning strikes looked and I'm going to redesign the animation sequence that reveals the Zenitsu character.
Instead of having lightning strikes I am going to
animate a polygon line with a trim-paths effect
running up/down the character outlining the major
shapes of the character.

The way this effect works is as follows: Each line
will have a designated CSS class that contains the
animated line of the polygon. Each one is unique.
These CSS classes are added to each respective div
element using a for loop.