Added CSS variable for timing of the outlines.
Also fixed a minor issue where the clip-path:polygon wasn't
defined properly for the first frame, thus resulting in
a solid white square (because there was no clipping the
first frame would just fill all of the space with white).

Will continue to work on this tomorrow. I have the basic idea
of the lightning strikes down. Once everything is finished
I'll come up with a way of looping through all of the pieces
in the proper order in JS and then get the timing right with
the intersection observer.

One problem I'm thinking about is how a user is going to know
to stop scrolling to view the animation. As it stands right
now, someone could easily just scroll down past the animation
and miss pieces of it. I'm thinking of having a rotating triangle
in the center of each section that will pause and animate
into the art piece for that respective section. This is going
to add a lot more work because I'll need to do this for each
section, but I think it'll make the project better.