Tree Base and Branch Animations Wokring
I figured out how to animate the tree base in first and
then have a branch animate in after. This way I can have
the base animate in and then have all the other branches
animate at once.

However, I feel like the final product of the animation
is too violent and choppy. I wanted this to look more
like a tree was yawning and stretching itself awake, but
this animation reminds me more of someone violently
chopping food in a kitchen or something. I think the
effect of the 5-tone brown shceme looks cool... but it's
absolutely not how I imagined it to be.

Thus, I'm going to redesign the base of the tree and have
it composed of fewer polygons. The polygons will slowly
grow into place and I think that will have a much better
effect than what we have now.