Created new Leaf Animation method I found the best way to animate the clumps of leaves in was to emulate a scaling animation from 0 to 100, but using a central origin point.
This way I can control the origin of the leaf bundle
to make it look like they're popping in one on top
of the other.

The problem with the opacity fade in is that I
intentionally am hiding rough sections of the
leaf bundles behind other bundles--and you can
clearly see them with an opacity animation. Using
a layered pop-in effect deals with this issue
and makes the animation look more organic than
a simple fade-in effect.

I'm going to stick with this animation for the
remainder of the section.