Finished Tree Animation The project is mostly done at this point. I still have to come up with unique animations for the first two sections.
The best way to go about this is to delay the animation
for a little more than 1 second. This way as the user
scrolls and reads the name of the section, the
intersection observer will trigger on a delay.

Without having the delay, the animation starts too
quickly and the user can't read the name of the
section.